- Implemented update and delete service for thread entity - Fixed authorization mechanism for services to use SessionService instead of interacting directly with flaskcontext - Fixed a bug in update service where update was not being processed correctly

+++ server/server/services/delete_service.py
@@ -7,7 +7,6 @@ import logging
 import server.exceptions as exceptions
 from server.database.filter import PrimitiveFilter
 from server.database.paging import PagingNoLimit
-from server.services.session import SessionService
 
 
 class DeleteService:
@@ -16,9 +15,9 @@ class DeleteService:
     Includes most of the business logic that is associated
     with deleting entities like users, threads etc.
     """
-    def __init__(self, repo, context):
+    def __init__(self, repo, session):
         self._repo = repo
-        self._context = context
+        self._session = session
 
     def deleteUserByKeyValues(self, keyValues):
         """
@@ -30,13 +29,13 @@ class DeleteService:
             result of operation
         """
         userId = keyValues.get('userId')
+        if userId is None:
+            logging.error('Failed to delete due to no Id specified in keyValues')
+            raise exceptions.IdNotSpecifiedError('Specify entity Id to perform delete')
         self._authorizeDelete(userId)
-        searchFilter = PrimitiveFilter.createFilter(dict(
-            operator='eq', field='userId', value=[ userId ]
-        ))
-        result = self._repo.deleteUser(searchFilter)
-
-        return dict(result=result)
+        searchFilter = self._createSearchFilterFromKeyValuesForField(keyValues, 'userId')
+        
+        return self._repo.deleteUser(searchFilter)
 
     def deletePostByKeyValues(self, keyValues):
         """
@@ -47,35 +46,76 @@ class DeleteService:
         Returns:
             result of operation
         """
-        postId = keyValues.get('postId')
-        self._authorizeDeletePost(postId)
-        searchFilter = PrimitiveFilter.createFilter(dict(
-            operator='eq', field='postId', value=[ postId ]
+        searchFilter = self._createSearchFilterFromKeyValuesForField(keyValues, 'postId')
+        self._authorizeDeletePost(searchFilter)
+        
+        return self._repo.deletePost(searchFilter)
+
+    def deleteThreadByKeyValues(self, keyValues):
+        """
+        Delete thread by passing criteria as keyValues
+        
+        Args:
+            keyValues(dict): criteria of thread to delete
+        Returns:
+            result of operation
+        """
+        searchFilter = self._createSearchFilterFromKeyValuesForField(keyValues, 'threadId')
+        self._authorizeDeleteThread(searchFilter)
+        
+        return self._repo.deleteThread(searchFilter)
+
+    def _createSearchFilterFromKeyValuesForField(self, keyValues, fieldname):
+        """
+        Creates searchFilter used to search entities from repo,
+        by extracting fieldname from keyvalues
+        
+        Args:
+            keyValues(dict):
+            fieldname(string): creates searchfilter using this fieldname
+        Returns:
+            PrimitiveFilter
+        """
+        fieldvalue = keyValues.get(fieldname, None)
+        if fieldvalue is None:
+            logging.error('Failed to delete due to no Id specified in keyValues')
+            raise exceptions.IdNotSpecifiedError('Specify entity Id to perform delete')
+        
+        return PrimitiveFilter.createFilter(dict(
+            operator='eq', field=fieldname, value=[ fieldvalue ]
         ))
-        result = self._repo.deletePost(searchFilter)
 
-        return dict(result=result)
-
-    def _authorizeDeletePost(self, postId):
+    def _authorizeDeletePost(self, searchFilter):
         """
         Authorizes current user to delete this post.
         
         Args:
-            postId(string): postId of post
+            searchFilter(PrimitiveFilter): used to search posts from repo
         Returns:
             None
         """
-        searchFilter = PrimitiveFilter.createFilter(dict(
-            operator='eq', field='postId', value=[ postId ]
-        ))
         posts = self._repo.searchPost(searchFilter, PagingNoLimit())['posts']
         
         for post in posts:
             self._authorizeDelete( post['userId'] )
 
+    def _authorizeDeleteThread(self, searchFilter):
+        """
+        Authorizes current user to delete this thread.
+        
+        Args:
+            searchFilter(PrimitiveFilter): used to search threads from repo
+        Returns:
+            None
+        """
+        threads = self._repo.searchThread(searchFilter, PagingNoLimit())['threads']
+        
+        for thread in threads:
+            self._authorizeDelete( thread['userId'] )
+
     def _authorizeDelete(self, userId):
         """
-        Authorizes delete operationby comparing resource owner with current user.
+        Authorizes delete operation by comparing resource owner with current user.
         Inteded to be used for cases like preventing user A from deleting posts created by user B.
         
         Args:
@@ -83,7 +123,7 @@ class DeleteService:
         Returns:
             None
         """
-        session_user = self._context.read_global(SessionService.SESSION_USER_KEY)
+        session_user = self._session.get_user()
         if (session_user is None) or (session_user['userId'] != userId):
             logging.error('Failed operation due to unauthorized user')
             raise exceptions.UnauthorizedError('Unauthorized action')
